pyproject.toml
# pyproject.toml
# Estif Bingo 24/7 - Python Project Configuration

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "estif-bingo-telegram-bot"
version = "3.0.0"
description = "Estif Bingo 24/7 - Telegram Bot for Bingo Game Management"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Estif Bingo Team", email = "support@estifbingo.com"}
]
license = {text = "Proprietary"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]

dependencies = [
    "python-telegram-bot>=20.7",
    "python-dotenv>=1.0.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "werkzeug>=3.0.1",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.1",
    "cryptography>=41.0.7",
    "requests>=2.31.0",
    "httpx>=0.25.2",
    "aiohttp>=3.9.1",
    "websockets>=12.0",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    "tzdata>=2023.3",
    "email-validator>=2.1.0",
    "phonenumbers>=8.13.26",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "orjson>=3.9.10",
    "cachetools>=5.3.1",
    "psutil>=5.9.6",
    "python-json-logger>=2.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
    "watchdog>=3.0.0",
    "ipython>=8.18.1",
    "debugpy>=1.8.0",
]

performance = [
    "uvloop>=0.19.0",
    "nest-asyncio>=1.5.8",
]

api = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
]

all = [
    "uvloop>=0.19.0",
    "nest-asyncio>=1.5.8",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/johnsonestiph13/estif-bingo-advanced"
Repository = "https://github.com/johnsonestiph13/estif-bingo-advanced"
Issues = "https://github.com/johnsonestiph13/estif-bingo-advanced/issues"

# ==================== TOOL CONFIGURATIONS ====================

[tool.black]
line-length = 100
target-version = ['py311']
exclude = '''
/(
    \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | node_modules
    | venv
    | __pycache__
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["venv/*", "__pycache__/*", "build/*", "dist/*"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
exclude = [
    "venv/",
    "__pycache__/",
    "build/",
    "dist/",
    "migrations/",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["bot"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
    "*/venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if False:",
]

# ==================== PDM CONFIGURATION (if using PDM) ====================
[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
]

# ==================== RUFF LINTER CONFIGURATION (optional) ====================
[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

# ==================== VIRTUAL ENVIRONMENT ====================
[tool.virtualenv]
create = true
download = true